genai/pipeline/functions/ingestor/requirements.txt
functions-framework==3.*
vertexai
google-cloud-aiplatform
langchain
pandas
duckdb==1.1.0
google-cloud-secret-manager
pinecone